# openapi.yaml
openapi: 3.0.0
info:
  title: AFOCE Accounting API
  description: |
    Complete accounting API for invoicing, expenses, and financial management.
    Supports multi-tenant workspaces with role-based access control.
  version: 1.0.0
  contact:
    name: AFOCE Support
    email: support@afoce.app
servers:
  - url: https://api.afoce.app
    description: Production server
  - url: http://localhost:3000
    description: Local development
security:
  - cookieAuth: []
paths:
...
  /api/invoices:
    get:
      summary: List invoices
      operationId: listInvoices
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, pending, paid, overdue, rejected]
        - name: clientId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of invoices
    post:
      summary: Create invoice
      operationId: createInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
      responses:
        "201":
          description: Invoice created
  /api/invoices/{id}/status:
    patch:
      summary: Update invoice status
      operationId: updateInvoiceStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
      responses:
        "200":
          description: Status updated
...
  /api/export:
    get:
      summary: Export data
      operationId: exportData
      parameters:
        - name: entity
          in: query
          required: true
          schema:
            type: string
            enum: [invoices, expenses, clients, bank_lines]
        - name: format
          in: query
          required: true
          schema:
            type: string
            enum: [csv, pdf, xlsx]
      responses:
        "200":
          description: Exported file
          content:
            text/csv:
              schema:
                type: string
            application/pdf:
              schema:
                type: string
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: supabase-auth-token
  schemas:
...
    CreateInvoiceRequest:
      type: object
      required:
        - client_id
        - bs_date
        - ad_date
        - amount
      properties:
        client_id:
          type: string
        bs_date:
          type: string
        ad_date:
          type: string
        amount:
          type: number
        due_days:
          type: integer
        status:
          type: string
...
    UpdateStatusRequest:
      type: object
      properties:
        status:
          type: string
        reason:
          type: string
...